+++ src/App.tsx
@@ -25,7 +25,7 @@ import DropdowndetailsPage from './components/DropdowndetailsPage';
 import AuthenticationCard from './components/AuthenticationCards';
 import ToastPage from './components/ToastPage';
 import AccordionDetails from './components/AccordionDetails';
-import ContactUsPage from './components/ContactUsPage';
+import ContactUsDetailsPage from './components/ContactUsDetailsPage';
 import PaginationDetails from './components/PaginationDetails';
 import TestimonialDetails from './components/TestimonialDetails';
 
@@ -64,7 +64,7 @@ const App: React.FC = () => {
         <Route path='/donate' element={<DonationPage />} />
 
         <Route path='/about' element={<AboutUsPage />} />
-        <Route path='/contact' element={<ContactUsPage />} />
+        <Route path='/contact-details' element={<ContactUsDetailsPage />} />
         <Route path='/pagination-details' element={<PaginationDetails />} />
         <Route path='/testimonial-details' element={<TestimonialDetails />} />
